feat: update GetMaxCapacityAsync parameters from integer to double

+++ Common/BusinessProviders/CompartmentProvider.cs
@@ -425,7 +425,7 @@ namespace Ferretto.Common.BusinessProviders
             return l;
         }
 
-        public async Task<int?> GetMaxCapacityAsync(int? width, int? height, int itemId)
+        public async Task<int?> GetMaxCapacityAsync(double? width, double? height, int itemId)
         {
             if (width.HasValue && height.HasValue)
             {
